Refactor Csv methods to new CsvHandler class
